+++ CHANGES.rst
@@ -2,8 +2,16 @@
 Changelog
 =========
 
-master
-======
+1.4.0 - 2017-03-05
+==================
+
+- ``MicroException`` renamed to ``DjangoMicroException``
+- using ``BASE_DIR/templates`` as default templates dir
+- fix detection application name
+- testing support
+
+1.3.0 - 2016-10-17
+==================
 
 - add support function-based commands
 
@@ -19,7 +27,7 @@ master
 - fix no module named %yourapp%
 
 1.1.0 - 2016-10-01
-===============
+==================
 
 - ``route`` and ``command`` may be used directly, not only as decorator
 - add support class-based views for ``route`` decorator